handlers/extraction_features/requirements.txt
azure-ai-formrecognizer
azure-servicebus
azure-storage-blob
python-dotenv